build(fuse): support static compilation for curvine-fuse
- Add `server` feature to curvine-common to make rocksdb, raft, flate2
  optional dependencies (only needed for master/worker nodes)
- Move RaftPeer and NodeId to conf module to avoid circular dependencies
- Add `static` feature to curvine-fuse to skip libfuse pkg-config check
- Fix musl libc compatibility issues in orpc and curvine-fuse:
  - Handle TMPFS_MAGIC type difference (musl: i64, glibc: u64)
  - Handle ioctl request param type difference (musl: i32, glibc: u64)
  - Handle request_code_read! macro return type difference

This enables building a fully static curvine-fuse binary using musl:
  cargo build --release --target x86_64-unknown-linux-musl \
    -p curvine-fuse --features static --no-default-features

+++ curvine-fuse/src/fuse_utils.rs
@@ -129,7 +129,8 @@ impl FuseUtils {
         {
             use nix;
             // request_code_read! macro is equivalent to linux _IOR macro
-            nix::request_code_read!(229, 0, 4)
+            // Cast to u64 to handle type differences between musl (i32) and glibc (u64)
+            nix::request_code_read!(229, 0, 4) as u64
         }
     }
